Gen AI-Day4/Task3/requirements.txt
crewai
streamlit
python-dotenv
langchain-openai
